+++ src/Symfony/Component/Mailer/Tests/Transport/Smtp/EsmtpTransportFactoryTest.php
@@ -38,30 +38,30 @@ class EsmtpTransportFactoryTest extends TransportFactoryTestCase
         $eventDispatcher = $this->getDispatcher();
         $logger = $this->getLogger();
 
-        $transport = new EsmtpTransport('localhost', 25, false, null, $eventDispatcher, $logger);
+        $transport = new EsmtpTransport('localhost', 25, false, $eventDispatcher, $logger);
 
         yield [
             new Dsn('smtp', 'localhost'),
             $transport,
         ];
 
-        $transport = new EsmtpTransport('example.com', 99, true, 'login', $eventDispatcher, $logger);
+        $transport = new EsmtpTransport('example.com', 99, true, $eventDispatcher, $logger);
         $transport->setUsername(self::USER);
         $transport->setPassword(self::PASSWORD);
 
         yield [
-            new Dsn('smtps', 'example.com', self::USER, self::PASSWORD, 99, ['auth_mode' => 'login']),
+            new Dsn('smtps', 'example.com', self::USER, self::PASSWORD, 99),
             $transport,
         ];
 
-        $transport = new EsmtpTransport('example.com', 465, true, null, $eventDispatcher, $logger);
+        $transport = new EsmtpTransport('example.com', 465, true, $eventDispatcher, $logger);
 
         yield [
             new Dsn('smtps', 'example.com'),
             $transport,
         ];
 
-        $transport = new EsmtpTransport('example.com', 465, true, null, $eventDispatcher, $logger);
+        $transport = new EsmtpTransport('example.com', 465, true, $eventDispatcher, $logger);
 
         yield [
             new Dsn('smtp', 'example.com', '', '', 465),